[Feat] - Add further domain objects and apply DDD strategies (#4)
* Started adding further Domain Entities and ValueObjects like Participants, Trainers, Rooms and Sessions.

* Added trainers, rooms, schedules and participants.

* Consolidated the models according to: aggregate roots, entities and value objects.

* Changed the eventual consistency to be async event based.

* Fixed UTs due to changes from Pipelinr Notifications to Spring Events.

* Fixed naming on UTs of deleting subscription.

* More eventual consistency on the subscription creation. Fixed respective UTs.

* Made changes to room to keep track of session IDs by date, since the max is daily.

* Started consolidating all aggregates, entities and value objects with all their business use cases. Refactored some things that felt wrong when mapping from storage.

* Added tests for trainers and sessions of gyms.

* Added further entities + further profiles and first cases of Rooms.

* Started adding session cases. Added trainer events and refactored session persistence to have just 1 category.

* Continued with adding session case.

* Refined the domain models so that session categories are part of the gym, since a gym might only support certain types of sessions. Started adding the session domain events handling.

* Added eventual consistency. Need to change Session Table start and end time to timestamp instead of date.

* Finished adding session case.

* Changed the start time and end time of session back to Timestampz instead of LocalTime.

* Fixed issue Participant Schedule was a Text instead of a JSONB on DB. Added case to make a session reservation.

* Fixed broken Unit Tests.

* Restructured class locations and added proper mappers to User class

* Added missing classes from previous commit.

* Fixed tests on Application.

* Added tests to profiles.

* Added tests to adding session trainer to gym event handler.

* Added further tests to the application layer.

* Added more tests to the application layer.

* Added tests to the adding session use cases.

* Added first Integration Tests for user registration with Tests Containers.

* Did the rest of the integration tests on the registration.

* Finished integration tests for Identity. Will try to generalize the TestContainers to avoid repetition between test classes.

* Added a test containers superclass for all integration tests. Moved the UTs under test.groovy.com.faroc.gymanager.unit

* Finished user profiles tests.

* Fixed unit tests imports.

* Refactored subscription command to Subscribe. Added tests to subscription cases.

* Made admin created method static.

* Fixed bug on the user update, the id was missing on the where clause.

* Fixed bug on admin repository updates. First tests for gym add.

* Finished gym integration tests.

* Added Spring Modulith, in specific its integration with events to store them and retry in case of failure.

* Made all event handlers @Service and added @ApplicationModuleListener, which is equivalent to @Async and @TransactionalEventListener.

* Solved some warnings.

---------

+++ gymanager.api/src/main/java/com/faroc/gymanager/application/subscriptions/commands/deletesubscription/DeleteSubscriptionHandler.java
@@ -1,29 +1,30 @@
 package com.faroc.gymanager.application.subscriptions.commands.deletesubscription;
 
 import an.awesome.pipelinr.Command;
-import an.awesome.pipelinr.Pipeline;
 import an.awesome.pipelinr.Voidy;
 import com.faroc.gymanager.application.admins.gateways.AdminsGateway;
 import com.faroc.gymanager.application.shared.exceptions.ResourceNotFoundException;
 import com.faroc.gymanager.application.subscriptions.gateways.SubscriptionsGateway;
 import com.faroc.gymanager.domain.admins.errors.AdminErrors;
+import com.faroc.gymanager.domain.shared.exceptions.UnexpectedException;
 import com.faroc.gymanager.domain.subscriptions.errors.SubscriptionErrors;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 @Component
 public class DeleteSubscriptionHandler implements Command.Handler<DeleteSubscriptionCommand, Voidy> {
-    private final Pipeline pipeline;
     private final AdminsGateway adminsGateway;
     private final SubscriptionsGateway subscriptionsGateway;
+    private final ApplicationEventPublisher eventsPublisher;
 
     public DeleteSubscriptionHandler(
-            Pipeline pipeline,
             AdminsGateway adminsGateway,
-            SubscriptionsGateway subscriptionsGateway) {
-        this.pipeline = pipeline;
+            SubscriptionsGateway subscriptionsGateway,
+            ApplicationEventPublisher eventsPublisher) {
         this.adminsGateway = adminsGateway;
         this.subscriptionsGateway = subscriptionsGateway;
+        this.eventsPublisher = eventsPublisher;
     }
 
     @Override
@@ -43,11 +44,18 @@ public class DeleteSubscriptionHandler implements Command.Handler<DeleteSubscrip
                         AdminErrors.NOT_FOUND
                 ));
 
-        admin.deleteSubscription(subscriptionId);
+        if (admin.getSubscriptionId() == null)
+            throw new UnexpectedException(
+                    AdminErrors.notFound(adminId),
+                    AdminErrors.SUBSCRIPTION_NOT_FOUND
+            );
 
+        admin.deleteSubscription(subscriptionId);
         adminsGateway.update(admin);
 
-        admin.getDomainEvents().forEach(event -> event.send(pipeline));
+        while (admin.hasDomainEvents()) {
+            eventsPublisher.publishEvent(admin.popEvent());
+        }
 
         return new Voidy();
     }
